pyproject.toml
[project]
name = "chatterbox-playground"
version = "0.1.0"
description = "CLI tool for text-to-speech synthesis using Chatterbox models"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "numpy<2.0",
  "chatterbox-tts",
  "ml-dtypes<0.5.0",
  "onnx<1.17.0",
  "resemble-perth==1.0.1",
  "setuptools",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu124", marker = "sys_platform == 'win32'" }]
torchaudio = [{ index = "pytorch-cu124", marker = "sys_platform == 'win32'" }]

[[tool.uv.index]]
name = "pytorch-cu124"
url  = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv]
no-build-isolation-package = ["pkuseg"]
